model/requirements.txt
tqdm
numpy
pymadcad[stl]
